+++ src/lib/Scenes/Home/__tests__/index-tests.tsx
@@ -4,6 +4,6 @@ import React from "react"
 import Home from "../index"
 
 it("has the correct number of tabs", () => {
-  const wrapper = shallow(<Home tracking={null} />)
+  const wrapper = shallow(<Home initialTab={0} isVisible={true} tracking={null} />)
   expect(wrapper.find(Tab)).toHaveLength(3)
 })